requirements.txt
streamlit==1.37.0
chromadb==0.5.5
sentence-transformers==2.7.0
python-dotenv==1.0.1
openai>=1.0,<3
tiktoken==0.7.0
pypdf==4.2.0
click==8.1.7
httpx<0.28
